- Agregue precio unitario en Carrito - Modifique la forma que el Nro de Comprobante pasa al Carrito y como se genera el Total en el Detalle Pedido

+++ TpProgramacion3-2C-Varela/Dominio/CarritoC.cs
@@ -9,7 +9,7 @@ namespace Dominio
     public class CarritoC
     {
         //public int ID { get; set; }
-        public int NROCOMPROBANTE { get; set; }
+        public Int64 NROCOMPROBANTE { get; set; }
         public Articulo ARTICULO { get; set; }
         public int CANTIDAD { get; set; }
         public decimal MONTO { get; set; }
